requirements.txt

djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
Pillow==9.0.0
django-debug-toolbar~=3.2.4
Django~=4.2.6